@@ -5,7 +5,7 @@ version: 0.12.0+21
 resolution: workspace
 
 environment:
-  sdk: '^3.9.2'
+  sdk: '^3.10.0'
 
 dependencies:
   cupertino_icons: ^1.0.8
@@ -17,7 +17,7 @@ dependencies:
     sdk: flutter
   freezed_annotation: ^3.1.0
   form_builder_validators: ^11.0.0
-  go_router: ^16.2.4
+  go_router: ^17.0.0
   intl: ^0.20.0
   json_annotation: ^4.9.0
   logging: ^1.2.0
