@@ -4,7 +4,7 @@ version: 1.0.0
 resolution: workspace
 
 environment:
-  sdk: ^3.9.2
+  sdk: ^3.10.0
 
 dependencies:
   freezed_annotation: ^3.1.0
