@@ -5,8 +5,8 @@ publish_to: none
 resolution: workspace
 
 environment:
-  sdk: ^3.9.2
-  flutter: ">=1.17.0"
+  sdk: ^3.10.0
+  flutter: ">=3.38.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
   theme_tailor_annotation: ^3.0.1
   slang: ^4.8.1
   slang_flutter: ^4.8.0
-  go_router: ^16.2.4
+  go_router: ^17.0.0
   intl: ^0.20.0
   nz_common:
     path: ../nz_common
