@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nz_flutter_core/nz_flutter_core.dart';
@@ -9,25 +25,31 @@ class HomeNavigationBlock extends StatelessWidget {
   Widget build(BuildContext context) {
     return AdaptiveListLimiter(
       child: Padding(
-        padding: const EdgeInsets.all(8),
+        padding: const .all(8),
         child: Column(
           children: [
             _NavigationCard(
               route: '/tasks',
               icon: Icons.format_list_bulleted_rounded,
-              title: t.tasks.list.title,
+              title: context.t.tasks.list.title,
+            ),
+            const SizedBox(height: 6),
+            _NavigationCard(
+              route: '/habits',
+              icon: Icons.published_with_changes_rounded,
+              title: context.t.habits.list.title,
             ),
             const SizedBox(height: 6),
             _NavigationCard(
               route: '/stats',
               icon: Icons.bar_chart_rounded,
-              title: t.stats.quickView.title,
+              title: context.t.stats.quickView.title,
             ),
             const SizedBox(height: 6),
             _NavigationCard(
               route: '/settings',
               icon: Icons.settings_rounded,
-              title: t.settings.list.title,
+              title: context.t.settings.list.title,
             ),
           ],
         ),
@@ -55,7 +77,7 @@ class _NavigationCard extends StatelessWidget {
         child: InkWell(
           onTap: () => context.push(route),
           child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16),
+            padding: const .symmetric(horizontal: 16),
             child: Row(
               children: [
                 Icon(
@@ -65,7 +87,7 @@ class _NavigationCard extends StatelessWidget {
                 const SizedBox(width: 10),
                 Text(
                   title,
-                  style: Theme.of(context).textTheme.titleLarge,
+                  style: context.theme.textTheme.titleLarge,
                 ),
               ],
             ),
