@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'dart:async';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -14,8 +30,8 @@ class QuickStatisticsNotifier extends Notifier<QuickStatisticsState> {
   }
 
   Future<void> loadDays([DateTime? start, DateTime? end]) async {
-    final rangeStart = start ?? DateTime.now().startOfWeek;
-    final rangeEnd = end ?? DateTime.now().endOfWeek;
+    final rangeStart = start ?? .now().startOfWeek;
+    final rangeEnd = end ?? .now().endOfWeek;
 
     final repository = ref.read(statsRepositoryProvider);
     final weeklyStats = await repository.getStatsByDays(rangeStart, rangeEnd);
