@@ -1,9 +1,26 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:drift/drift.dart';
 import 'package:not_zero_app/src/features/stats/models/tasks_counting_data.dart';
 import 'package:nz_base_models/nz_base_models.dart';
+import 'package:nz_common/nz_common.dart';
 import 'package:nz_drift/nz_drift.dart';
 
-class TasksStatsLocalService {
+class TasksStatsLocalService implements BaseService {
   const TasksStatsLocalService(this._db);
 
   final NotZeroDatabase _db;
@@ -12,21 +29,20 @@ class TasksStatsLocalService {
     DateTime? startPeriod,
     DateTime? endPeriod,
   }) async {
-    const enumConverter = EnumIndexConverter(TaskImportance.values);
     return _db.transaction(() async {
       return TasksCountingData(
         notImportant: await _completedInPeriodByImportance(
-          importance: enumConverter.toSql(TaskImportance.notImportant),
+          importance: .notImportant,
           startPeriod: startPeriod,
           endPeriod: endPeriod,
         ),
         normal: await _completedInPeriodByImportance(
-          importance: enumConverter.toSql(TaskImportance.normal),
+          importance: .normal,
           startPeriod: startPeriod,
           endPeriod: endPeriod,
         ),
         important: await _completedInPeriodByImportance(
-          importance: enumConverter.toSql(TaskImportance.important),
+          importance: .important,
           startPeriod: startPeriod,
           endPeriod: endPeriod,
         ),
@@ -39,7 +55,7 @@ class TasksStatsLocalService {
   }
 
   Future<int> _completedInPeriodByImportance({
-    required int importance,
+    required TaskImportance importance,
     DateTime? startPeriod,
     DateTime? endPeriod,
   }) async {
@@ -47,7 +63,7 @@ class TasksStatsLocalService {
       ..addColumns({countAll()})
       ..where(
         _db.tasksTable.completedAt.inPeriod(startPeriod, endPeriod) &
-            _db.tasksTable.importance.equals(importance),
+            _db.tasksTable.importance.equalsValue(importance),
       );
     final result = await query.getSingleOrNull();
     return result?.read(countAll()) ?? 0;
