@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:collection/collection.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:not_zero_app/src/features/actions_bus/di.dart';
@@ -22,10 +38,15 @@ final tasksRepositoryProvider = Provider<TasksRepository>((ref) {
 });
 
 final specificTaskStreamProvider = FutureProvider.autoDispose
-    .family<Task?, String>((ref, taskId) async {
-      // TODO(uSlashVlad): Doesn't work if a task isn't from the main list :(
-      final mainList = await ref.watch(tasksMainListNotifier.future);
-      return mainList.firstWhereOrNull((task) => task.id == taskId);
+    .family<Task?, String>((
+      ref,
+      taskId,
+    ) {
+      return ref.watch(
+        tasksMainListNotifier.selectAsync(
+          (list) => list.firstWhereOrNull((task) => task.id == taskId),
+        ),
+      );
     });
 
 final tasksFiltersNotifier = NotifierProvider.autoDispose(
