@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
@@ -20,18 +36,18 @@ class TasksEmptyListPlaceholder extends ConsumerWidget {
 
     final theme = Theme.of(context);
     return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
+      mainAxisAlignment: .center,
       children: [
         const SizedBox(height: 24),
         Text(
-          t.tasks.list.empty.icon,
+          context.t.tasks.list.empty.icon,
           style: theme.textTheme.displayLarge?.copyWith(fontSize: 48),
         ),
         const SizedBox(height: 16),
         Text(
           hasFilters
-              ? t.tasks.list.empty.titleFilters
-              : t.tasks.list.empty.title,
+              ? context.t.tasks.list.empty.titleFilters
+              : context.t.tasks.list.empty.title,
           style: theme.textTheme.titleLarge,
         ),
         const SizedBox(height: 8),
@@ -45,8 +61,8 @@ class TasksEmptyListPlaceholder extends ConsumerWidget {
           },
           child: Text(
             hasFilters
-                ? t.tasks.list.empty.buttonFilters
-                : t.tasks.list.empty.button,
+                ? context.t.tasks.list.empty.buttonFilters
+                : context.t.tasks.list.empty.button,
           ),
         ),
         const SizedBox(height: 24),
