@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'dart:io';
 
 import 'package:drift/drift.dart';
@@ -5,19 +21,29 @@ import 'package:nz_base_models/nz_base_models.dart';
 import 'package:nz_drift/src/converters/date_converter.dart';
 import 'package:nz_drift/src/database/native_database.dart';
 import 'package:nz_drift/src/migrations/migrations.dart';
+import 'package:nz_drift/src/tables/habits_table.dart';
 import 'package:nz_drift/src/tables/tags_table.dart';
 import 'package:nz_drift/src/tables/tasks_table.dart';
 
 part 'db.g.dart';
 
-@DriftDatabase(tables: [TasksTable, TagsTable, TasksTagEntries])
+@DriftDatabase(
+  tables: [
+    TasksTable,
+    TagsTable,
+    TasksTagEntries,
+    HabitsTable,
+    HabitCompletionsTable,
+    HabitsTagEntries,
+  ],
+)
 class NotZeroDatabase extends _$NotZeroDatabase {
   NotZeroDatabase([QueryExecutor? e]) : super(e ?? openDriftDatabase());
 
   NotZeroDatabase.memory() : super(openDriftDatabase(permanent: false));
 
   @override
-  int get schemaVersion => 6;
+  int get schemaVersion => 7;
 
   @override
   MigrationStrategy get migration {
