@@ -215,12 +215,6 @@ class $TasksTableTable extends TasksTable
         DriftSqlType.string,
         data['${effectivePrefix}title'],
       )!,
-      importance: $TasksTableTable.$converterimportance.fromSql(
-        attachedDatabase.typeMapping.read(
-          DriftSqlType.int,
-          data['${effectivePrefix}importance'],
-        )!,
-      ),
       createdAt: attachedDatabase.typeMapping.read(
         DriftSqlType.dateTime,
         data['${effectivePrefix}created_at'],
@@ -229,6 +223,12 @@ class $TasksTableTable extends TasksTable
         DriftSqlType.string,
         data['${effectivePrefix}description'],
       )!,
+      importance: $TasksTableTable.$converterimportance.fromSql(
+        attachedDatabase.typeMapping.read(
+          DriftSqlType.int,
+          data['${effectivePrefix}importance'],
+        )!,
+      ),
       modifiedAt: attachedDatabase.typeMapping.read(
         DriftSqlType.dateTime,
         data['${effectivePrefix}modified_at'],
@@ -896,12 +896,820 @@ class TasksTagEntriesCompanion extends UpdateCompanion<TasksTagEntry> {
   }
 }
 
+class $HabitsTableTable extends HabitsTable
+    with TableInfo<$HabitsTableTable, Habit> {
+  @override
+  final GeneratedDatabase attachedDatabase;
+  final String? _alias;
+  $HabitsTableTable(this.attachedDatabase, [this._alias]);
+  static const VerificationMeta _idMeta = const VerificationMeta('id');
+  @override
+  late final GeneratedColumn<String> id = GeneratedColumn<String>(
+    'id',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+  );
+  static const VerificationMeta _titleMeta = const VerificationMeta('title');
+  @override
+  late final GeneratedColumn<String> title = GeneratedColumn<String>(
+    'title',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+  );
+  static const VerificationMeta _descriptionMeta = const VerificationMeta(
+    'description',
+  );
+  @override
+  late final GeneratedColumn<String> description = GeneratedColumn<String>(
+    'description',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: false,
+    defaultValue: const Constant(''),
+  );
+  static const VerificationMeta _createdAtMeta = const VerificationMeta(
+    'createdAt',
+  );
+  @override
+  late final GeneratedColumn<DateTime> createdAt = GeneratedColumn<DateTime>(
+    'created_at',
+    aliasedName,
+    false,
+    type: DriftSqlType.dateTime,
+    requiredDuringInsert: false,
+    defaultValue: currentDateAndTime,
+  );
+  static const VerificationMeta _modifiedAtMeta = const VerificationMeta(
+    'modifiedAt',
+  );
+  @override
+  late final GeneratedColumn<DateTime> modifiedAt = GeneratedColumn<DateTime>(
+    'modified_at',
+    aliasedName,
+    true,
+    type: DriftSqlType.dateTime,
+    requiredDuringInsert: false,
+  );
+  @override
+  late final GeneratedColumnWithTypeConverter<TaskImportance, int> importance =
+      GeneratedColumn<int>(
+        'importance',
+        aliasedName,
+        false,
+        type: DriftSqlType.int,
+        requiredDuringInsert: true,
+      ).withConverter<TaskImportance>($HabitsTableTable.$converterimportance);
+  @override
+  List<GeneratedColumn> get $columns => [
+    id,
+    title,
+    description,
+    createdAt,
+    modifiedAt,
+    importance,
+  ];
+  @override
+  String get aliasedName => _alias ?? actualTableName;
+  @override
+  String get actualTableName => $name;
+  static const String $name = 'habits_table';
+  @override
+  VerificationContext validateIntegrity(
+    Insertable<Habit> instance, {
+    bool isInserting = false,
+  }) {
+    final context = VerificationContext();
+    final data = instance.toColumns(true);
+    if (data.containsKey('id')) {
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
+    } else if (isInserting) {
+      context.missing(_idMeta);
+    }
+    if (data.containsKey('title')) {
+      context.handle(
+        _titleMeta,
+        title.isAcceptableOrUnknown(data['title']!, _titleMeta),
+      );
+    } else if (isInserting) {
+      context.missing(_titleMeta);
+    }
+    if (data.containsKey('description')) {
+      context.handle(
+        _descriptionMeta,
+        description.isAcceptableOrUnknown(
+          data['description']!,
+          _descriptionMeta,
+        ),
+      );
+    }
+    if (data.containsKey('created_at')) {
+      context.handle(
+        _createdAtMeta,
+        createdAt.isAcceptableOrUnknown(data['created_at']!, _createdAtMeta),
+      );
+    }
+    if (data.containsKey('modified_at')) {
+      context.handle(
+        _modifiedAtMeta,
+        modifiedAt.isAcceptableOrUnknown(data['modified_at']!, _modifiedAtMeta),
+      );
+    }
+    return context;
+  }
+
+  @override
+  Set<GeneratedColumn> get $primaryKey => {id};
+  @override
+  Habit map(Map<String, dynamic> data, {String? tablePrefix}) {
+    final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : '';
+    return Habit(
+      id: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}id'],
+      )!,
+      title: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}title'],
+      )!,
+      createdAt: attachedDatabase.typeMapping.read(
+        DriftSqlType.dateTime,
+        data['${effectivePrefix}created_at'],
+      )!,
+      description: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}description'],
+      )!,
+      importance: $HabitsTableTable.$converterimportance.fromSql(
+        attachedDatabase.typeMapping.read(
+          DriftSqlType.int,
+          data['${effectivePrefix}importance'],
+        )!,
+      ),
+      modifiedAt: attachedDatabase.typeMapping.read(
+        DriftSqlType.dateTime,
+        data['${effectivePrefix}modified_at'],
+      ),
+    );
+  }
+
+  @override
+  $HabitsTableTable createAlias(String alias) {
+    return $HabitsTableTable(attachedDatabase, alias);
+  }
+
+  static JsonTypeConverter2<TaskImportance, int, int> $converterimportance =
+      const EnumIndexConverter<TaskImportance>(TaskImportance.values);
+}
+
+class HabitsTableCompanion extends UpdateCompanion<Habit> {
+  final Value<String> id;
+  final Value<String> title;
+  final Value<String> description;
+  final Value<DateTime> createdAt;
+  final Value<DateTime?> modifiedAt;
+  final Value<TaskImportance> importance;
+  final Value<int> rowid;
+  const HabitsTableCompanion({
+    this.id = const Value.absent(),
+    this.title = const Value.absent(),
+    this.description = const Value.absent(),
+    this.createdAt = const Value.absent(),
+    this.modifiedAt = const Value.absent(),
+    this.importance = const Value.absent(),
+    this.rowid = const Value.absent(),
+  });
+  HabitsTableCompanion.insert({
+    required String id,
+    required String title,
+    this.description = const Value.absent(),
+    this.createdAt = const Value.absent(),
+    this.modifiedAt = const Value.absent(),
+    required TaskImportance importance,
+    this.rowid = const Value.absent(),
+  }) : id = Value(id),
+       title = Value(title),
+       importance = Value(importance);
+  static Insertable<Habit> custom({
+    Expression<String>? id,
+    Expression<String>? title,
+    Expression<String>? description,
+    Expression<DateTime>? createdAt,
+    Expression<DateTime>? modifiedAt,
+    Expression<int>? importance,
+    Expression<int>? rowid,
+  }) {
+    return RawValuesInsertable({
+      if (id != null) 'id': id,
+      if (title != null) 'title': title,
+      if (description != null) 'description': description,
+      if (createdAt != null) 'created_at': createdAt,
+      if (modifiedAt != null) 'modified_at': modifiedAt,
+      if (importance != null) 'importance': importance,
+      if (rowid != null) 'rowid': rowid,
+    });
+  }
+
+  HabitsTableCompanion copyWith({
+    Value<String>? id,
+    Value<String>? title,
+    Value<String>? description,
+    Value<DateTime>? createdAt,
+    Value<DateTime?>? modifiedAt,
+    Value<TaskImportance>? importance,
+    Value<int>? rowid,
+  }) {
+    return HabitsTableCompanion(
+      id: id ?? this.id,
+      title: title ?? this.title,
+      description: description ?? this.description,
+      createdAt: createdAt ?? this.createdAt,
+      modifiedAt: modifiedAt ?? this.modifiedAt,
+      importance: importance ?? this.importance,
+      rowid: rowid ?? this.rowid,
+    );
+  }
+
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    final map = <String, Expression>{};
+    if (id.present) {
+      map['id'] = Variable<String>(id.value);
+    }
+    if (title.present) {
+      map['title'] = Variable<String>(title.value);
+    }
+    if (description.present) {
+      map['description'] = Variable<String>(description.value);
+    }
+    if (createdAt.present) {
+      map['created_at'] = Variable<DateTime>(createdAt.value);
+    }
+    if (modifiedAt.present) {
+      map['modified_at'] = Variable<DateTime>(modifiedAt.value);
+    }
+    if (importance.present) {
+      map['importance'] = Variable<int>(
+        $HabitsTableTable.$converterimportance.toSql(importance.value),
+      );
+    }
+    if (rowid.present) {
+      map['rowid'] = Variable<int>(rowid.value);
+    }
+    return map;
+  }
+
+  @override
+  String toString() {
+    return (StringBuffer('HabitsTableCompanion(')
+          ..write('id: $id, ')
+          ..write('title: $title, ')
+          ..write('description: $description, ')
+          ..write('createdAt: $createdAt, ')
+          ..write('modifiedAt: $modifiedAt, ')
+          ..write('importance: $importance, ')
+          ..write('rowid: $rowid')
+          ..write(')'))
+        .toString();
+  }
+}
+
+class _$HabitInsertable implements Insertable<Habit> {
+  Habit _object;
+  _$HabitInsertable(this._object);
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    return HabitsTableCompanion(
+      id: Value(_object.id),
+      title: Value(_object.title),
+      description: Value(_object.description),
+      createdAt: Value(_object.createdAt),
+      modifiedAt: Value(_object.modifiedAt),
+      importance: Value(_object.importance),
+    ).toColumns(false);
+  }
+}
+
+extension HabitToInsertable on Habit {
+  _$HabitInsertable toInsertable() {
+    return _$HabitInsertable(this);
+  }
+}
+
+class $HabitCompletionsTableTable extends HabitCompletionsTable
+    with TableInfo<$HabitCompletionsTableTable, HabitCompletion> {
+  @override
+  final GeneratedDatabase attachedDatabase;
+  final String? _alias;
+  $HabitCompletionsTableTable(this.attachedDatabase, [this._alias]);
+  static const VerificationMeta _idMeta = const VerificationMeta('id');
+  @override
+  late final GeneratedColumn<String> id = GeneratedColumn<String>(
+    'id',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+  );
+  static const VerificationMeta _habitIdMeta = const VerificationMeta(
+    'habitId',
+  );
+  @override
+  late final GeneratedColumn<String> habitId = GeneratedColumn<String>(
+    'habit_id',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+    defaultConstraints: GeneratedColumn.constraintIsAlways(
+      'REFERENCES habits_table (id)',
+    ),
+  );
+  @override
+  late final GeneratedColumnWithTypeConverter<HabitCompletionType, int> type =
+      GeneratedColumn<int>(
+        'type',
+        aliasedName,
+        false,
+        type: DriftSqlType.int,
+        requiredDuringInsert: true,
+      ).withConverter<HabitCompletionType>(
+        $HabitCompletionsTableTable.$convertertype,
+      );
+  @override
+  late final GeneratedColumnWithTypeConverter<DateTime, String> completedDate =
+      GeneratedColumn<String>(
+        'completed_date',
+        aliasedName,
+        false,
+        type: DriftSqlType.string,
+        requiredDuringInsert: true,
+      ).withConverter<DateTime>(
+        $HabitCompletionsTableTable.$convertercompletedDate,
+      );
+  static const VerificationMeta _streakCountMeta = const VerificationMeta(
+    'streakCount',
+  );
+  @override
+  late final GeneratedColumn<int> streakCount = GeneratedColumn<int>(
+    'streak_count',
+    aliasedName,
+    false,
+    type: DriftSqlType.int,
+    requiredDuringInsert: false,
+    defaultValue: const Constant(1),
+  );
+  @override
+  List<GeneratedColumn> get $columns => [
+    id,
+    habitId,
+    type,
+    completedDate,
+    streakCount,
+  ];
+  @override
+  String get aliasedName => _alias ?? actualTableName;
+  @override
+  String get actualTableName => $name;
+  static const String $name = 'habit_completions_table';
+  @override
+  VerificationContext validateIntegrity(
+    Insertable<HabitCompletion> instance, {
+    bool isInserting = false,
+  }) {
+    final context = VerificationContext();
+    final data = instance.toColumns(true);
+    if (data.containsKey('id')) {
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
+    } else if (isInserting) {
+      context.missing(_idMeta);
+    }
+    if (data.containsKey('habit_id')) {
+      context.handle(
+        _habitIdMeta,
+        habitId.isAcceptableOrUnknown(data['habit_id']!, _habitIdMeta),
+      );
+    } else if (isInserting) {
+      context.missing(_habitIdMeta);
+    }
+    if (data.containsKey('streak_count')) {
+      context.handle(
+        _streakCountMeta,
+        streakCount.isAcceptableOrUnknown(
+          data['streak_count']!,
+          _streakCountMeta,
+        ),
+      );
+    }
+    return context;
+  }
+
+  @override
+  Set<GeneratedColumn> get $primaryKey => {id};
+  @override
+  List<Set<GeneratedColumn>> get uniqueKeys => [
+    {habitId, completedDate},
+  ];
+  @override
+  HabitCompletion map(Map<String, dynamic> data, {String? tablePrefix}) {
+    final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : '';
+    return HabitCompletion(
+      id: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}id'],
+      )!,
+      habitId: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}habit_id'],
+      )!,
+      type: $HabitCompletionsTableTable.$convertertype.fromSql(
+        attachedDatabase.typeMapping.read(
+          DriftSqlType.int,
+          data['${effectivePrefix}type'],
+        )!,
+      ),
+      completedDate: $HabitCompletionsTableTable.$convertercompletedDate
+          .fromSql(
+            attachedDatabase.typeMapping.read(
+              DriftSqlType.string,
+              data['${effectivePrefix}completed_date'],
+            )!,
+          ),
+      streakCount: attachedDatabase.typeMapping.read(
+        DriftSqlType.int,
+        data['${effectivePrefix}streak_count'],
+      )!,
+    );
+  }
+
+  @override
+  $HabitCompletionsTableTable createAlias(String alias) {
+    return $HabitCompletionsTableTable(attachedDatabase, alias);
+  }
+
+  static JsonTypeConverter2<HabitCompletionType, int, int> $convertertype =
+      const EnumIndexConverter<HabitCompletionType>(HabitCompletionType.values);
+  static TypeConverter<DateTime, String> $convertercompletedDate =
+      const DateConverter();
+}
+
+class HabitCompletionsTableCompanion extends UpdateCompanion<HabitCompletion> {
+  final Value<String> id;
+  final Value<String> habitId;
+  final Value<HabitCompletionType> type;
+  final Value<DateTime> completedDate;
+  final Value<int> streakCount;
+  final Value<int> rowid;
+  const HabitCompletionsTableCompanion({
+    this.id = const Value.absent(),
+    this.habitId = const Value.absent(),
+    this.type = const Value.absent(),
+    this.completedDate = const Value.absent(),
+    this.streakCount = const Value.absent(),
+    this.rowid = const Value.absent(),
+  });
+  HabitCompletionsTableCompanion.insert({
+    required String id,
+    required String habitId,
+    required HabitCompletionType type,
+    required DateTime completedDate,
+    this.streakCount = const Value.absent(),
+    this.rowid = const Value.absent(),
+  }) : id = Value(id),
+       habitId = Value(habitId),
+       type = Value(type),
+       completedDate = Value(completedDate);
+  static Insertable<HabitCompletion> custom({
+    Expression<String>? id,
+    Expression<String>? habitId,
+    Expression<int>? type,
+    Expression<String>? completedDate,
+    Expression<int>? streakCount,
+    Expression<int>? rowid,
+  }) {
+    return RawValuesInsertable({
+      if (id != null) 'id': id,
+      if (habitId != null) 'habit_id': habitId,
+      if (type != null) 'type': type,
+      if (completedDate != null) 'completed_date': completedDate,
+      if (streakCount != null) 'streak_count': streakCount,
+      if (rowid != null) 'rowid': rowid,
+    });
+  }
+
+  HabitCompletionsTableCompanion copyWith({
+    Value<String>? id,
+    Value<String>? habitId,
+    Value<HabitCompletionType>? type,
+    Value<DateTime>? completedDate,
+    Value<int>? streakCount,
+    Value<int>? rowid,
+  }) {
+    return HabitCompletionsTableCompanion(
+      id: id ?? this.id,
+      habitId: habitId ?? this.habitId,
+      type: type ?? this.type,
+      completedDate: completedDate ?? this.completedDate,
+      streakCount: streakCount ?? this.streakCount,
+      rowid: rowid ?? this.rowid,
+    );
+  }
+
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    final map = <String, Expression>{};
+    if (id.present) {
+      map['id'] = Variable<String>(id.value);
+    }
+    if (habitId.present) {
+      map['habit_id'] = Variable<String>(habitId.value);
+    }
+    if (type.present) {
+      map['type'] = Variable<int>(
+        $HabitCompletionsTableTable.$convertertype.toSql(type.value),
+      );
+    }
+    if (completedDate.present) {
+      map['completed_date'] = Variable<String>(
+        $HabitCompletionsTableTable.$convertercompletedDate.toSql(
+          completedDate.value,
+        ),
+      );
+    }
+    if (streakCount.present) {
+      map['streak_count'] = Variable<int>(streakCount.value);
+    }
+    if (rowid.present) {
+      map['rowid'] = Variable<int>(rowid.value);
+    }
+    return map;
+  }
+
+  @override
+  String toString() {
+    return (StringBuffer('HabitCompletionsTableCompanion(')
+          ..write('id: $id, ')
+          ..write('habitId: $habitId, ')
+          ..write('type: $type, ')
+          ..write('completedDate: $completedDate, ')
+          ..write('streakCount: $streakCount, ')
+          ..write('rowid: $rowid')
+          ..write(')'))
+        .toString();
+  }
+}
+
+class _$HabitCompletionInsertable implements Insertable<HabitCompletion> {
+  HabitCompletion _object;
+  _$HabitCompletionInsertable(this._object);
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    return HabitCompletionsTableCompanion(
+      id: Value(_object.id),
+      habitId: Value(_object.habitId),
+      type: Value(_object.type),
+      completedDate: Value(_object.completedDate),
+      streakCount: Value(_object.streakCount),
+    ).toColumns(false);
+  }
+}
+
+extension HabitCompletionToInsertable on HabitCompletion {
+  _$HabitCompletionInsertable toInsertable() {
+    return _$HabitCompletionInsertable(this);
+  }
+}
+
+class $HabitsTagEntriesTable extends HabitsTagEntries
+    with TableInfo<$HabitsTagEntriesTable, HabitsTagEntry> {
+  @override
+  final GeneratedDatabase attachedDatabase;
+  final String? _alias;
+  $HabitsTagEntriesTable(this.attachedDatabase, [this._alias]);
+  static const VerificationMeta _habitMeta = const VerificationMeta('habit');
+  @override
+  late final GeneratedColumn<String> habit = GeneratedColumn<String>(
+    'habit',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+    defaultConstraints: GeneratedColumn.constraintIsAlways(
+      'REFERENCES habits_table (id)',
+    ),
+  );
+  static const VerificationMeta _tagMeta = const VerificationMeta('tag');
+  @override
+  late final GeneratedColumn<String> tag = GeneratedColumn<String>(
+    'tag',
+    aliasedName,
+    false,
+    type: DriftSqlType.string,
+    requiredDuringInsert: true,
+    defaultConstraints: GeneratedColumn.constraintIsAlways(
+      'REFERENCES tags_table (id)',
+    ),
+  );
+  @override
+  List<GeneratedColumn> get $columns => [habit, tag];
+  @override
+  String get aliasedName => _alias ?? actualTableName;
+  @override
+  String get actualTableName => $name;
+  static const String $name = 'habits_tag_entries';
+  @override
+  VerificationContext validateIntegrity(
+    Insertable<HabitsTagEntry> instance, {
+    bool isInserting = false,
+  }) {
+    final context = VerificationContext();
+    final data = instance.toColumns(true);
+    if (data.containsKey('habit')) {
+      context.handle(
+        _habitMeta,
+        habit.isAcceptableOrUnknown(data['habit']!, _habitMeta),
+      );
+    } else if (isInserting) {
+      context.missing(_habitMeta);
+    }
+    if (data.containsKey('tag')) {
+      context.handle(
+        _tagMeta,
+        tag.isAcceptableOrUnknown(data['tag']!, _tagMeta),
+      );
+    } else if (isInserting) {
+      context.missing(_tagMeta);
+    }
+    return context;
+  }
+
+  @override
+  Set<GeneratedColumn> get $primaryKey => {habit, tag};
+  @override
+  HabitsTagEntry map(Map<String, dynamic> data, {String? tablePrefix}) {
+    final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : '';
+    return HabitsTagEntry(
+      habit: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}habit'],
+      )!,
+      tag: attachedDatabase.typeMapping.read(
+        DriftSqlType.string,
+        data['${effectivePrefix}tag'],
+      )!,
+    );
+  }
+
+  @override
+  $HabitsTagEntriesTable createAlias(String alias) {
+    return $HabitsTagEntriesTable(attachedDatabase, alias);
+  }
+}
+
+class HabitsTagEntry extends DataClass implements Insertable<HabitsTagEntry> {
+  final String habit;
+  final String tag;
+  const HabitsTagEntry({required this.habit, required this.tag});
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    final map = <String, Expression>{};
+    map['habit'] = Variable<String>(habit);
+    map['tag'] = Variable<String>(tag);
+    return map;
+  }
+
+  HabitsTagEntriesCompanion toCompanion(bool nullToAbsent) {
+    return HabitsTagEntriesCompanion(habit: Value(habit), tag: Value(tag));
+  }
+
+  factory HabitsTagEntry.fromJson(
+    Map<String, dynamic> json, {
+    ValueSerializer? serializer,
+  }) {
+    serializer ??= driftRuntimeOptions.defaultSerializer;
+    return HabitsTagEntry(
+      habit: serializer.fromJson<String>(json['habit']),
+      tag: serializer.fromJson<String>(json['tag']),
+    );
+  }
+  @override
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
+    serializer ??= driftRuntimeOptions.defaultSerializer;
+    return <String, dynamic>{
+      'habit': serializer.toJson<String>(habit),
+      'tag': serializer.toJson<String>(tag),
+    };
+  }
+
+  HabitsTagEntry copyWith({String? habit, String? tag}) =>
+      HabitsTagEntry(habit: habit ?? this.habit, tag: tag ?? this.tag);
+  HabitsTagEntry copyWithCompanion(HabitsTagEntriesCompanion data) {
+    return HabitsTagEntry(
+      habit: data.habit.present ? data.habit.value : this.habit,
+      tag: data.tag.present ? data.tag.value : this.tag,
+    );
+  }
+
+  @override
+  String toString() {
+    return (StringBuffer('HabitsTagEntry(')
+          ..write('habit: $habit, ')
+          ..write('tag: $tag')
+          ..write(')'))
+        .toString();
+  }
+
+  @override
+  int get hashCode => Object.hash(habit, tag);
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      (other is HabitsTagEntry &&
+          other.habit == this.habit &&
+          other.tag == this.tag);
+}
+
+class HabitsTagEntriesCompanion extends UpdateCompanion<HabitsTagEntry> {
+  final Value<String> habit;
+  final Value<String> tag;
+  final Value<int> rowid;
+  const HabitsTagEntriesCompanion({
+    this.habit = const Value.absent(),
+    this.tag = const Value.absent(),
+    this.rowid = const Value.absent(),
+  });
+  HabitsTagEntriesCompanion.insert({
+    required String habit,
+    required String tag,
+    this.rowid = const Value.absent(),
+  }) : habit = Value(habit),
+       tag = Value(tag);
+  static Insertable<HabitsTagEntry> custom({
+    Expression<String>? habit,
+    Expression<String>? tag,
+    Expression<int>? rowid,
+  }) {
+    return RawValuesInsertable({
+      if (habit != null) 'habit': habit,
+      if (tag != null) 'tag': tag,
+      if (rowid != null) 'rowid': rowid,
+    });
+  }
+
+  HabitsTagEntriesCompanion copyWith({
+    Value<String>? habit,
+    Value<String>? tag,
+    Value<int>? rowid,
+  }) {
+    return HabitsTagEntriesCompanion(
+      habit: habit ?? this.habit,
+      tag: tag ?? this.tag,
+      rowid: rowid ?? this.rowid,
+    );
+  }
+
+  @override
+  Map<String, Expression> toColumns(bool nullToAbsent) {
+    final map = <String, Expression>{};
+    if (habit.present) {
+      map['habit'] = Variable<String>(habit.value);
+    }
+    if (tag.present) {
+      map['tag'] = Variable<String>(tag.value);
+    }
+    if (rowid.present) {
+      map['rowid'] = Variable<int>(rowid.value);
+    }
+    return map;
+  }
+
+  @override
+  String toString() {
+    return (StringBuffer('HabitsTagEntriesCompanion(')
+          ..write('habit: $habit, ')
+          ..write('tag: $tag, ')
+          ..write('rowid: $rowid')
+          ..write(')'))
+        .toString();
+  }
+}
+
 abstract class _$NotZeroDatabase extends GeneratedDatabase {
   _$NotZeroDatabase(QueryExecutor e) : super(e);
   $NotZeroDatabaseManager get managers => $NotZeroDatabaseManager(this);
   late final $TasksTableTable tasksTable = $TasksTableTable(this);
   late final $TagsTableTable tagsTable = $TagsTableTable(this);
   late final $TasksTagEntriesTable tasksTagEntries = $TasksTagEntriesTable(
+    this,
+  );
+  late final $HabitsTableTable habitsTable = $HabitsTableTable(this);
+  late final $HabitCompletionsTableTable habitCompletionsTable =
+      $HabitCompletionsTableTable(this);
+  late final $HabitsTagEntriesTable habitsTagEntries = $HabitsTagEntriesTable(
     this,
   );
   @override
@@ -912,6 +1720,9 @@ abstract class _$NotZeroDatabase extends GeneratedDatabase {
     tasksTable,
     tagsTable,
     tasksTagEntries,
+    habitsTable,
+    habitCompletionsTable,
+    habitsTagEntries,
   ];
 }
 
@@ -1378,6 +2189,30 @@ final class $$TagsTableTableReferences
       manager.$state.copyWith(prefetchedData: cache),
     );
   }
+
+  static MultiTypedResultKey<$HabitsTagEntriesTable, List<HabitsTagEntry>>
+  _habitsTagEntriesRefsTable(_$NotZeroDatabase db) =>
+      MultiTypedResultKey.fromTable(
+        db.habitsTagEntries,
+        aliasName: $_aliasNameGenerator(
+          db.tagsTable.id,
+          db.habitsTagEntries.tag,
+        ),
+      );
+
+  $$HabitsTagEntriesTableProcessedTableManager get habitsTagEntriesRefs {
+    final manager = $$HabitsTagEntriesTableTableManager(
+      $_db,
+      $_db.habitsTagEntries,
+    ).filter((f) => f.tag.id.sqlEquals($_itemColumn<String>('id')!));
+
+    final cache = $_typedResult.readTableOrNull(
+      _habitsTagEntriesRefsTable($_db),
+    );
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: cache),
+    );
+  }
 }
 
 class $$TagsTableTableFilterComposer
@@ -1425,6 +2260,31 @@ class $$TagsTableTableFilterComposer
           }) => $$TasksTagEntriesTableFilterComposer(
             $db: $db,
             $table: $db.tasksTagEntries,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return f(composer);
+  }
+
+  Expression<bool> habitsTagEntriesRefs(
+    Expression<bool> Function($$HabitsTagEntriesTableFilterComposer f) f,
+  ) {
+    final $$HabitsTagEntriesTableFilterComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.id,
+      referencedTable: $db.habitsTagEntries,
+      getReferencedColumn: (t) => t.tag,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTagEntriesTableFilterComposer(
+            $db: $db,
+            $table: $db.habitsTagEntries,
             $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
             joinBuilder: joinBuilder,
             $removeJoinBuilderFromRootComposer:
@@ -1512,6 +2372,31 @@ class $$TagsTableTableAnnotationComposer
     );
     return f(composer);
   }
+
+  Expression<T> habitsTagEntriesRefs<T extends Object>(
+    Expression<T> Function($$HabitsTagEntriesTableAnnotationComposer a) f,
+  ) {
+    final $$HabitsTagEntriesTableAnnotationComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.id,
+      referencedTable: $db.habitsTagEntries,
+      getReferencedColumn: (t) => t.tag,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTagEntriesTableAnnotationComposer(
+            $db: $db,
+            $table: $db.habitsTagEntries,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return f(composer);
+  }
 }
 
 class $$TagsTableTableTableManager
@@ -1527,7 +2412,10 @@ class $$TagsTableTableTableManager
           $$TagsTableTableUpdateCompanionBuilder,
           (ItemTag, $$TagsTableTableReferences),
           ItemTag,
-          PrefetchHooks Function({bool tasksTagEntriesRefs})
+          PrefetchHooks Function({
+            bool tasksTagEntriesRefs,
+            bool habitsTagEntriesRefs,
+          })
         > {
   $$TagsTableTableTableManager(_$NotZeroDatabase db, $TagsTableTable table)
     : super(
@@ -1576,38 +2464,63 @@ class $$TagsTableTableTableManager
                 ),
               )
               .toList(),
-          prefetchHooksCallback: ({tasksTagEntriesRefs = false}) {
-            return PrefetchHooks(
-              db: db,
-              explicitlyWatchedTables: [
-                if (tasksTagEntriesRefs) db.tasksTagEntries,
-              ],
-              addJoins: null,
-              getPrefetchedDataCallback: (items) async {
-                return [
-                  if (tasksTagEntriesRefs)
-                    await $_getPrefetchedData<
-                      ItemTag,
-                      $TagsTableTable,
-                      TasksTagEntry
-                    >(
-                      currentTable: table,
-                      referencedTable: $$TagsTableTableReferences
-                          ._tasksTagEntriesRefsTable(db),
-                      managerFromTypedResult: (p0) =>
-                          $$TagsTableTableReferences(
-                            db,
-                            table,
-                            p0,
-                          ).tasksTagEntriesRefs,
-                      referencedItemsForCurrentItem: (item, referencedItems) =>
-                          referencedItems.where((e) => e.tag == item.id),
-                      typedResults: items,
-                    ),
-                ];
+          prefetchHooksCallback:
+              ({tasksTagEntriesRefs = false, habitsTagEntriesRefs = false}) {
+                return PrefetchHooks(
+                  db: db,
+                  explicitlyWatchedTables: [
+                    if (tasksTagEntriesRefs) db.tasksTagEntries,
+                    if (habitsTagEntriesRefs) db.habitsTagEntries,
+                  ],
+                  addJoins: null,
+                  getPrefetchedDataCallback: (items) async {
+                    return [
+                      if (tasksTagEntriesRefs)
+                        await $_getPrefetchedData<
+                          ItemTag,
+                          $TagsTableTable,
+                          TasksTagEntry
+                        >(
+                          currentTable: table,
+                          referencedTable: $$TagsTableTableReferences
+                              ._tasksTagEntriesRefsTable(db),
+                          managerFromTypedResult: (p0) =>
+                              $$TagsTableTableReferences(
+                                db,
+                                table,
+                                p0,
+                              ).tasksTagEntriesRefs,
+                          referencedItemsForCurrentItem:
+                              (item, referencedItems) => referencedItems.where(
+                                (e) => e.tag == item.id,
+                              ),
+                          typedResults: items,
+                        ),
+                      if (habitsTagEntriesRefs)
+                        await $_getPrefetchedData<
+                          ItemTag,
+                          $TagsTableTable,
+                          HabitsTagEntry
+                        >(
+                          currentTable: table,
+                          referencedTable: $$TagsTableTableReferences
+                              ._habitsTagEntriesRefsTable(db),
+                          managerFromTypedResult: (p0) =>
+                              $$TagsTableTableReferences(
+                                db,
+                                table,
+                                p0,
+                              ).habitsTagEntriesRefs,
+                          referencedItemsForCurrentItem:
+                              (item, referencedItems) => referencedItems.where(
+                                (e) => e.tag == item.id,
+                              ),
+                          typedResults: items,
+                        ),
+                    ];
+                  },
+                );
               },
-            );
-          },
         ),
       );
 }
@@ -1624,7 +2537,10 @@ typedef $$TagsTableTableProcessedTableManager =
       $$TagsTableTableUpdateCompanionBuilder,
       (ItemTag, $$TagsTableTableReferences),
       ItemTag,
-      PrefetchHooks Function({bool tasksTagEntriesRefs})
+      PrefetchHooks Function({
+        bool tasksTagEntriesRefs,
+        bool habitsTagEntriesRefs,
+      })
     >;
 typedef $$TasksTagEntriesTableCreateCompanionBuilder =
     TasksTagEntriesCompanion Function({
@@ -1990,6 +2906,1170 @@ typedef $$TasksTagEntriesTableProcessedTableManager =
       TasksTagEntry,
       PrefetchHooks Function({bool task, bool tag})
     >;
+typedef $$HabitsTableTableCreateCompanionBuilder =
+    HabitsTableCompanion Function({
+      required String id,
+      required String title,
+      Value<String> description,
+      Value<DateTime> createdAt,
+      Value<DateTime?> modifiedAt,
+      required TaskImportance importance,
+      Value<int> rowid,
+    });
+typedef $$HabitsTableTableUpdateCompanionBuilder =
+    HabitsTableCompanion Function({
+      Value<String> id,
+      Value<String> title,
+      Value<String> description,
+      Value<DateTime> createdAt,
+      Value<DateTime?> modifiedAt,
+      Value<TaskImportance> importance,
+      Value<int> rowid,
+    });
+
+final class $$HabitsTableTableReferences
+    extends BaseReferences<_$NotZeroDatabase, $HabitsTableTable, Habit> {
+  $$HabitsTableTableReferences(super.$_db, super.$_table, super.$_typedResult);
+
+  static MultiTypedResultKey<$HabitCompletionsTableTable, List<HabitCompletion>>
+  _habitCompletionsTableRefsTable(_$NotZeroDatabase db) =>
+      MultiTypedResultKey.fromTable(
+        db.habitCompletionsTable,
+        aliasName: $_aliasNameGenerator(
+          db.habitsTable.id,
+          db.habitCompletionsTable.habitId,
+        ),
+      );
+
+  $$HabitCompletionsTableTableProcessedTableManager
+  get habitCompletionsTableRefs {
+    final manager = $$HabitCompletionsTableTableTableManager(
+      $_db,
+      $_db.habitCompletionsTable,
+    ).filter((f) => f.habitId.id.sqlEquals($_itemColumn<String>('id')!));
+
+    final cache = $_typedResult.readTableOrNull(
+      _habitCompletionsTableRefsTable($_db),
+    );
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: cache),
+    );
+  }
+
+  static MultiTypedResultKey<$HabitsTagEntriesTable, List<HabitsTagEntry>>
+  _habitsTagEntriesRefsTable(_$NotZeroDatabase db) =>
+      MultiTypedResultKey.fromTable(
+        db.habitsTagEntries,
+        aliasName: $_aliasNameGenerator(
+          db.habitsTable.id,
+          db.habitsTagEntries.habit,
+        ),
+      );
+
+  $$HabitsTagEntriesTableProcessedTableManager get habitsTagEntriesRefs {
+    final manager = $$HabitsTagEntriesTableTableManager(
+      $_db,
+      $_db.habitsTagEntries,
+    ).filter((f) => f.habit.id.sqlEquals($_itemColumn<String>('id')!));
+
+    final cache = $_typedResult.readTableOrNull(
+      _habitsTagEntriesRefsTable($_db),
+    );
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: cache),
+    );
+  }
+}
+
+class $$HabitsTableTableFilterComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTableTable> {
+  $$HabitsTableTableFilterComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnFilters<String> get id => $composableBuilder(
+    column: $table.id,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnFilters<String> get title => $composableBuilder(
+    column: $table.title,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnFilters<String> get description => $composableBuilder(
+    column: $table.description,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnFilters<DateTime> get createdAt => $composableBuilder(
+    column: $table.createdAt,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnFilters<DateTime> get modifiedAt => $composableBuilder(
+    column: $table.modifiedAt,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnWithTypeConverterFilters<TaskImportance, TaskImportance, int>
+  get importance => $composableBuilder(
+    column: $table.importance,
+    builder: (column) => ColumnWithTypeConverterFilters(column),
+  );
+
+  Expression<bool> habitCompletionsTableRefs(
+    Expression<bool> Function($$HabitCompletionsTableTableFilterComposer f) f,
+  ) {
+    final $$HabitCompletionsTableTableFilterComposer composer =
+        $composerBuilder(
+          composer: this,
+          getCurrentColumn: (t) => t.id,
+          referencedTable: $db.habitCompletionsTable,
+          getReferencedColumn: (t) => t.habitId,
+          builder:
+              (
+                joinBuilder, {
+                $addJoinBuilderToRootComposer,
+                $removeJoinBuilderFromRootComposer,
+              }) => $$HabitCompletionsTableTableFilterComposer(
+                $db: $db,
+                $table: $db.habitCompletionsTable,
+                $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+                joinBuilder: joinBuilder,
+                $removeJoinBuilderFromRootComposer:
+                    $removeJoinBuilderFromRootComposer,
+              ),
+        );
+    return f(composer);
+  }
+
+  Expression<bool> habitsTagEntriesRefs(
+    Expression<bool> Function($$HabitsTagEntriesTableFilterComposer f) f,
+  ) {
+    final $$HabitsTagEntriesTableFilterComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.id,
+      referencedTable: $db.habitsTagEntries,
+      getReferencedColumn: (t) => t.habit,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTagEntriesTableFilterComposer(
+            $db: $db,
+            $table: $db.habitsTagEntries,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return f(composer);
+  }
+}
+
+class $$HabitsTableTableOrderingComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTableTable> {
+  $$HabitsTableTableOrderingComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnOrderings<String> get id => $composableBuilder(
+    column: $table.id,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<String> get title => $composableBuilder(
+    column: $table.title,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<String> get description => $composableBuilder(
+    column: $table.description,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<DateTime> get createdAt => $composableBuilder(
+    column: $table.createdAt,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<DateTime> get modifiedAt => $composableBuilder(
+    column: $table.modifiedAt,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<int> get importance => $composableBuilder(
+    column: $table.importance,
+    builder: (column) => ColumnOrderings(column),
+  );
+}
+
+class $$HabitsTableTableAnnotationComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTableTable> {
+  $$HabitsTableTableAnnotationComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  GeneratedColumn<String> get id =>
+      $composableBuilder(column: $table.id, builder: (column) => column);
+
+  GeneratedColumn<String> get title =>
+      $composableBuilder(column: $table.title, builder: (column) => column);
+
+  GeneratedColumn<String> get description => $composableBuilder(
+    column: $table.description,
+    builder: (column) => column,
+  );
+
+  GeneratedColumn<DateTime> get createdAt =>
+      $composableBuilder(column: $table.createdAt, builder: (column) => column);
+
+  GeneratedColumn<DateTime> get modifiedAt => $composableBuilder(
+    column: $table.modifiedAt,
+    builder: (column) => column,
+  );
+
+  GeneratedColumnWithTypeConverter<TaskImportance, int> get importance =>
+      $composableBuilder(
+        column: $table.importance,
+        builder: (column) => column,
+      );
+
+  Expression<T> habitCompletionsTableRefs<T extends Object>(
+    Expression<T> Function($$HabitCompletionsTableTableAnnotationComposer a) f,
+  ) {
+    final $$HabitCompletionsTableTableAnnotationComposer composer =
+        $composerBuilder(
+          composer: this,
+          getCurrentColumn: (t) => t.id,
+          referencedTable: $db.habitCompletionsTable,
+          getReferencedColumn: (t) => t.habitId,
+          builder:
+              (
+                joinBuilder, {
+                $addJoinBuilderToRootComposer,
+                $removeJoinBuilderFromRootComposer,
+              }) => $$HabitCompletionsTableTableAnnotationComposer(
+                $db: $db,
+                $table: $db.habitCompletionsTable,
+                $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+                joinBuilder: joinBuilder,
+                $removeJoinBuilderFromRootComposer:
+                    $removeJoinBuilderFromRootComposer,
+              ),
+        );
+    return f(composer);
+  }
+
+  Expression<T> habitsTagEntriesRefs<T extends Object>(
+    Expression<T> Function($$HabitsTagEntriesTableAnnotationComposer a) f,
+  ) {
+    final $$HabitsTagEntriesTableAnnotationComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.id,
+      referencedTable: $db.habitsTagEntries,
+      getReferencedColumn: (t) => t.habit,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTagEntriesTableAnnotationComposer(
+            $db: $db,
+            $table: $db.habitsTagEntries,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return f(composer);
+  }
+}
+
+class $$HabitsTableTableTableManager
+    extends
+        RootTableManager<
+          _$NotZeroDatabase,
+          $HabitsTableTable,
+          Habit,
+          $$HabitsTableTableFilterComposer,
+          $$HabitsTableTableOrderingComposer,
+          $$HabitsTableTableAnnotationComposer,
+          $$HabitsTableTableCreateCompanionBuilder,
+          $$HabitsTableTableUpdateCompanionBuilder,
+          (Habit, $$HabitsTableTableReferences),
+          Habit,
+          PrefetchHooks Function({
+            bool habitCompletionsTableRefs,
+            bool habitsTagEntriesRefs,
+          })
+        > {
+  $$HabitsTableTableTableManager(_$NotZeroDatabase db, $HabitsTableTable table)
+    : super(
+        TableManagerState(
+          db: db,
+          table: table,
+          createFilteringComposer: () =>
+              $$HabitsTableTableFilterComposer($db: db, $table: table),
+          createOrderingComposer: () =>
+              $$HabitsTableTableOrderingComposer($db: db, $table: table),
+          createComputedFieldComposer: () =>
+              $$HabitsTableTableAnnotationComposer($db: db, $table: table),
+          updateCompanionCallback:
+              ({
+                Value<String> id = const Value.absent(),
+                Value<String> title = const Value.absent(),
+                Value<String> description = const Value.absent(),
+                Value<DateTime> createdAt = const Value.absent(),
+                Value<DateTime?> modifiedAt = const Value.absent(),
+                Value<TaskImportance> importance = const Value.absent(),
+                Value<int> rowid = const Value.absent(),
+              }) => HabitsTableCompanion(
+                id: id,
+                title: title,
+                description: description,
+                createdAt: createdAt,
+                modifiedAt: modifiedAt,
+                importance: importance,
+                rowid: rowid,
+              ),
+          createCompanionCallback:
+              ({
+                required String id,
+                required String title,
+                Value<String> description = const Value.absent(),
+                Value<DateTime> createdAt = const Value.absent(),
+                Value<DateTime?> modifiedAt = const Value.absent(),
+                required TaskImportance importance,
+                Value<int> rowid = const Value.absent(),
+              }) => HabitsTableCompanion.insert(
+                id: id,
+                title: title,
+                description: description,
+                createdAt: createdAt,
+                modifiedAt: modifiedAt,
+                importance: importance,
+                rowid: rowid,
+              ),
+          withReferenceMapper: (p0) => p0
+              .map(
+                (e) => (
+                  e.readTable(table),
+                  $$HabitsTableTableReferences(db, table, e),
+                ),
+              )
+              .toList(),
+          prefetchHooksCallback:
+              ({
+                habitCompletionsTableRefs = false,
+                habitsTagEntriesRefs = false,
+              }) {
+                return PrefetchHooks(
+                  db: db,
+                  explicitlyWatchedTables: [
+                    if (habitCompletionsTableRefs) db.habitCompletionsTable,
+                    if (habitsTagEntriesRefs) db.habitsTagEntries,
+                  ],
+                  addJoins: null,
+                  getPrefetchedDataCallback: (items) async {
+                    return [
+                      if (habitCompletionsTableRefs)
+                        await $_getPrefetchedData<
+                          Habit,
+                          $HabitsTableTable,
+                          HabitCompletion
+                        >(
+                          currentTable: table,
+                          referencedTable: $$HabitsTableTableReferences
+                              ._habitCompletionsTableRefsTable(db),
+                          managerFromTypedResult: (p0) =>
+                              $$HabitsTableTableReferences(
+                                db,
+                                table,
+                                p0,
+                              ).habitCompletionsTableRefs,
+                          referencedItemsForCurrentItem:
+                              (item, referencedItems) => referencedItems.where(
+                                (e) => e.habitId == item.id,
+                              ),
+                          typedResults: items,
+                        ),
+                      if (habitsTagEntriesRefs)
+                        await $_getPrefetchedData<
+                          Habit,
+                          $HabitsTableTable,
+                          HabitsTagEntry
+                        >(
+                          currentTable: table,
+                          referencedTable: $$HabitsTableTableReferences
+                              ._habitsTagEntriesRefsTable(db),
+                          managerFromTypedResult: (p0) =>
+                              $$HabitsTableTableReferences(
+                                db,
+                                table,
+                                p0,
+                              ).habitsTagEntriesRefs,
+                          referencedItemsForCurrentItem:
+                              (item, referencedItems) => referencedItems.where(
+                                (e) => e.habit == item.id,
+                              ),
+                          typedResults: items,
+                        ),
+                    ];
+                  },
+                );
+              },
+        ),
+      );
+}
+
+typedef $$HabitsTableTableProcessedTableManager =
+    ProcessedTableManager<
+      _$NotZeroDatabase,
+      $HabitsTableTable,
+      Habit,
+      $$HabitsTableTableFilterComposer,
+      $$HabitsTableTableOrderingComposer,
+      $$HabitsTableTableAnnotationComposer,
+      $$HabitsTableTableCreateCompanionBuilder,
+      $$HabitsTableTableUpdateCompanionBuilder,
+      (Habit, $$HabitsTableTableReferences),
+      Habit,
+      PrefetchHooks Function({
+        bool habitCompletionsTableRefs,
+        bool habitsTagEntriesRefs,
+      })
+    >;
+typedef $$HabitCompletionsTableTableCreateCompanionBuilder =
+    HabitCompletionsTableCompanion Function({
+      required String id,
+      required String habitId,
+      required HabitCompletionType type,
+      required DateTime completedDate,
+      Value<int> streakCount,
+      Value<int> rowid,
+    });
+typedef $$HabitCompletionsTableTableUpdateCompanionBuilder =
+    HabitCompletionsTableCompanion Function({
+      Value<String> id,
+      Value<String> habitId,
+      Value<HabitCompletionType> type,
+      Value<DateTime> completedDate,
+      Value<int> streakCount,
+      Value<int> rowid,
+    });
+
+final class $$HabitCompletionsTableTableReferences
+    extends
+        BaseReferences<
+          _$NotZeroDatabase,
+          $HabitCompletionsTableTable,
+          HabitCompletion
+        > {
+  $$HabitCompletionsTableTableReferences(
+    super.$_db,
+    super.$_table,
+    super.$_typedResult,
+  );
+
+  static $HabitsTableTable _habitIdTable(_$NotZeroDatabase db) =>
+      db.habitsTable.createAlias(
+        $_aliasNameGenerator(
+          db.habitCompletionsTable.habitId,
+          db.habitsTable.id,
+        ),
+      );
+
+  $$HabitsTableTableProcessedTableManager get habitId {
+    final $_column = $_itemColumn<String>('habit_id')!;
+
+    final manager = $$HabitsTableTableTableManager(
+      $_db,
+      $_db.habitsTable,
+    ).filter((f) => f.id.sqlEquals($_column));
+    final item = $_typedResult.readTableOrNull(_habitIdTable($_db));
+    if (item == null) return manager;
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: [item]),
+    );
+  }
+}
+
+class $$HabitCompletionsTableTableFilterComposer
+    extends Composer<_$NotZeroDatabase, $HabitCompletionsTableTable> {
+  $$HabitCompletionsTableTableFilterComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnFilters<String> get id => $composableBuilder(
+    column: $table.id,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  ColumnWithTypeConverterFilters<HabitCompletionType, HabitCompletionType, int>
+  get type => $composableBuilder(
+    column: $table.type,
+    builder: (column) => ColumnWithTypeConverterFilters(column),
+  );
+
+  ColumnWithTypeConverterFilters<DateTime, DateTime, String>
+  get completedDate => $composableBuilder(
+    column: $table.completedDate,
+    builder: (column) => ColumnWithTypeConverterFilters(column),
+  );
+
+  ColumnFilters<int> get streakCount => $composableBuilder(
+    column: $table.streakCount,
+    builder: (column) => ColumnFilters(column),
+  );
+
+  $$HabitsTableTableFilterComposer get habitId {
+    final $$HabitsTableTableFilterComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habitId,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableFilterComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitCompletionsTableTableOrderingComposer
+    extends Composer<_$NotZeroDatabase, $HabitCompletionsTableTable> {
+  $$HabitCompletionsTableTableOrderingComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnOrderings<String> get id => $composableBuilder(
+    column: $table.id,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<int> get type => $composableBuilder(
+    column: $table.type,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<String> get completedDate => $composableBuilder(
+    column: $table.completedDate,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  ColumnOrderings<int> get streakCount => $composableBuilder(
+    column: $table.streakCount,
+    builder: (column) => ColumnOrderings(column),
+  );
+
+  $$HabitsTableTableOrderingComposer get habitId {
+    final $$HabitsTableTableOrderingComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habitId,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableOrderingComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitCompletionsTableTableAnnotationComposer
+    extends Composer<_$NotZeroDatabase, $HabitCompletionsTableTable> {
+  $$HabitCompletionsTableTableAnnotationComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  GeneratedColumn<String> get id =>
+      $composableBuilder(column: $table.id, builder: (column) => column);
+
+  GeneratedColumnWithTypeConverter<HabitCompletionType, int> get type =>
+      $composableBuilder(column: $table.type, builder: (column) => column);
+
+  GeneratedColumnWithTypeConverter<DateTime, String> get completedDate =>
+      $composableBuilder(
+        column: $table.completedDate,
+        builder: (column) => column,
+      );
+
+  GeneratedColumn<int> get streakCount => $composableBuilder(
+    column: $table.streakCount,
+    builder: (column) => column,
+  );
+
+  $$HabitsTableTableAnnotationComposer get habitId {
+    final $$HabitsTableTableAnnotationComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habitId,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableAnnotationComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitCompletionsTableTableTableManager
+    extends
+        RootTableManager<
+          _$NotZeroDatabase,
+          $HabitCompletionsTableTable,
+          HabitCompletion,
+          $$HabitCompletionsTableTableFilterComposer,
+          $$HabitCompletionsTableTableOrderingComposer,
+          $$HabitCompletionsTableTableAnnotationComposer,
+          $$HabitCompletionsTableTableCreateCompanionBuilder,
+          $$HabitCompletionsTableTableUpdateCompanionBuilder,
+          (HabitCompletion, $$HabitCompletionsTableTableReferences),
+          HabitCompletion,
+          PrefetchHooks Function({bool habitId})
+        > {
+  $$HabitCompletionsTableTableTableManager(
+    _$NotZeroDatabase db,
+    $HabitCompletionsTableTable table,
+  ) : super(
+        TableManagerState(
+          db: db,
+          table: table,
+          createFilteringComposer: () =>
+              $$HabitCompletionsTableTableFilterComposer(
+                $db: db,
+                $table: table,
+              ),
+          createOrderingComposer: () =>
+              $$HabitCompletionsTableTableOrderingComposer(
+                $db: db,
+                $table: table,
+              ),
+          createComputedFieldComposer: () =>
+              $$HabitCompletionsTableTableAnnotationComposer(
+                $db: db,
+                $table: table,
+              ),
+          updateCompanionCallback:
+              ({
+                Value<String> id = const Value.absent(),
+                Value<String> habitId = const Value.absent(),
+                Value<HabitCompletionType> type = const Value.absent(),
+                Value<DateTime> completedDate = const Value.absent(),
+                Value<int> streakCount = const Value.absent(),
+                Value<int> rowid = const Value.absent(),
+              }) => HabitCompletionsTableCompanion(
+                id: id,
+                habitId: habitId,
+                type: type,
+                completedDate: completedDate,
+                streakCount: streakCount,
+                rowid: rowid,
+              ),
+          createCompanionCallback:
+              ({
+                required String id,
+                required String habitId,
+                required HabitCompletionType type,
+                required DateTime completedDate,
+                Value<int> streakCount = const Value.absent(),
+                Value<int> rowid = const Value.absent(),
+              }) => HabitCompletionsTableCompanion.insert(
+                id: id,
+                habitId: habitId,
+                type: type,
+                completedDate: completedDate,
+                streakCount: streakCount,
+                rowid: rowid,
+              ),
+          withReferenceMapper: (p0) => p0
+              .map(
+                (e) => (
+                  e.readTable(table),
+                  $$HabitCompletionsTableTableReferences(db, table, e),
+                ),
+              )
+              .toList(),
+          prefetchHooksCallback: ({habitId = false}) {
+            return PrefetchHooks(
+              db: db,
+              explicitlyWatchedTables: [],
+              addJoins:
+                  <
+                    T extends TableManagerState<
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic
+                    >
+                  >(state) {
+                    if (habitId) {
+                      state =
+                          state.withJoin(
+                                currentTable: table,
+                                currentColumn: table.habitId,
+                                referencedTable:
+                                    $$HabitCompletionsTableTableReferences
+                                        ._habitIdTable(db),
+                                referencedColumn:
+                                    $$HabitCompletionsTableTableReferences
+                                        ._habitIdTable(db)
+                                        .id,
+                              )
+                              as T;
+                    }
+
+                    return state;
+                  },
+              getPrefetchedDataCallback: (items) async {
+                return [];
+              },
+            );
+          },
+        ),
+      );
+}
+
+typedef $$HabitCompletionsTableTableProcessedTableManager =
+    ProcessedTableManager<
+      _$NotZeroDatabase,
+      $HabitCompletionsTableTable,
+      HabitCompletion,
+      $$HabitCompletionsTableTableFilterComposer,
+      $$HabitCompletionsTableTableOrderingComposer,
+      $$HabitCompletionsTableTableAnnotationComposer,
+      $$HabitCompletionsTableTableCreateCompanionBuilder,
+      $$HabitCompletionsTableTableUpdateCompanionBuilder,
+      (HabitCompletion, $$HabitCompletionsTableTableReferences),
+      HabitCompletion,
+      PrefetchHooks Function({bool habitId})
+    >;
+typedef $$HabitsTagEntriesTableCreateCompanionBuilder =
+    HabitsTagEntriesCompanion Function({
+      required String habit,
+      required String tag,
+      Value<int> rowid,
+    });
+typedef $$HabitsTagEntriesTableUpdateCompanionBuilder =
+    HabitsTagEntriesCompanion Function({
+      Value<String> habit,
+      Value<String> tag,
+      Value<int> rowid,
+    });
+
+final class $$HabitsTagEntriesTableReferences
+    extends
+        BaseReferences<
+          _$NotZeroDatabase,
+          $HabitsTagEntriesTable,
+          HabitsTagEntry
+        > {
+  $$HabitsTagEntriesTableReferences(
+    super.$_db,
+    super.$_table,
+    super.$_typedResult,
+  );
+
+  static $HabitsTableTable _habitTable(_$NotZeroDatabase db) =>
+      db.habitsTable.createAlias(
+        $_aliasNameGenerator(db.habitsTagEntries.habit, db.habitsTable.id),
+      );
+
+  $$HabitsTableTableProcessedTableManager get habit {
+    final $_column = $_itemColumn<String>('habit')!;
+
+    final manager = $$HabitsTableTableTableManager(
+      $_db,
+      $_db.habitsTable,
+    ).filter((f) => f.id.sqlEquals($_column));
+    final item = $_typedResult.readTableOrNull(_habitTable($_db));
+    if (item == null) return manager;
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: [item]),
+    );
+  }
+
+  static $TagsTableTable _tagTable(_$NotZeroDatabase db) =>
+      db.tagsTable.createAlias(
+        $_aliasNameGenerator(db.habitsTagEntries.tag, db.tagsTable.id),
+      );
+
+  $$TagsTableTableProcessedTableManager get tag {
+    final $_column = $_itemColumn<String>('tag')!;
+
+    final manager = $$TagsTableTableTableManager(
+      $_db,
+      $_db.tagsTable,
+    ).filter((f) => f.id.sqlEquals($_column));
+    final item = $_typedResult.readTableOrNull(_tagTable($_db));
+    if (item == null) return manager;
+    return ProcessedTableManager(
+      manager.$state.copyWith(prefetchedData: [item]),
+    );
+  }
+}
+
+class $$HabitsTagEntriesTableFilterComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTagEntriesTable> {
+  $$HabitsTagEntriesTableFilterComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  $$HabitsTableTableFilterComposer get habit {
+    final $$HabitsTableTableFilterComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habit,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableFilterComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+
+  $$TagsTableTableFilterComposer get tag {
+    final $$TagsTableTableFilterComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.tag,
+      referencedTable: $db.tagsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$TagsTableTableFilterComposer(
+            $db: $db,
+            $table: $db.tagsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitsTagEntriesTableOrderingComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTagEntriesTable> {
+  $$HabitsTagEntriesTableOrderingComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  $$HabitsTableTableOrderingComposer get habit {
+    final $$HabitsTableTableOrderingComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habit,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableOrderingComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+
+  $$TagsTableTableOrderingComposer get tag {
+    final $$TagsTableTableOrderingComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.tag,
+      referencedTable: $db.tagsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$TagsTableTableOrderingComposer(
+            $db: $db,
+            $table: $db.tagsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitsTagEntriesTableAnnotationComposer
+    extends Composer<_$NotZeroDatabase, $HabitsTagEntriesTable> {
+  $$HabitsTagEntriesTableAnnotationComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  $$HabitsTableTableAnnotationComposer get habit {
+    final $$HabitsTableTableAnnotationComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.habit,
+      referencedTable: $db.habitsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$HabitsTableTableAnnotationComposer(
+            $db: $db,
+            $table: $db.habitsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+
+  $$TagsTableTableAnnotationComposer get tag {
+    final $$TagsTableTableAnnotationComposer composer = $composerBuilder(
+      composer: this,
+      getCurrentColumn: (t) => t.tag,
+      referencedTable: $db.tagsTable,
+      getReferencedColumn: (t) => t.id,
+      builder:
+          (
+            joinBuilder, {
+            $addJoinBuilderToRootComposer,
+            $removeJoinBuilderFromRootComposer,
+          }) => $$TagsTableTableAnnotationComposer(
+            $db: $db,
+            $table: $db.tagsTable,
+            $addJoinBuilderToRootComposer: $addJoinBuilderToRootComposer,
+            joinBuilder: joinBuilder,
+            $removeJoinBuilderFromRootComposer:
+                $removeJoinBuilderFromRootComposer,
+          ),
+    );
+    return composer;
+  }
+}
+
+class $$HabitsTagEntriesTableTableManager
+    extends
+        RootTableManager<
+          _$NotZeroDatabase,
+          $HabitsTagEntriesTable,
+          HabitsTagEntry,
+          $$HabitsTagEntriesTableFilterComposer,
+          $$HabitsTagEntriesTableOrderingComposer,
+          $$HabitsTagEntriesTableAnnotationComposer,
+          $$HabitsTagEntriesTableCreateCompanionBuilder,
+          $$HabitsTagEntriesTableUpdateCompanionBuilder,
+          (HabitsTagEntry, $$HabitsTagEntriesTableReferences),
+          HabitsTagEntry,
+          PrefetchHooks Function({bool habit, bool tag})
+        > {
+  $$HabitsTagEntriesTableTableManager(
+    _$NotZeroDatabase db,
+    $HabitsTagEntriesTable table,
+  ) : super(
+        TableManagerState(
+          db: db,
+          table: table,
+          createFilteringComposer: () =>
+              $$HabitsTagEntriesTableFilterComposer($db: db, $table: table),
+          createOrderingComposer: () =>
+              $$HabitsTagEntriesTableOrderingComposer($db: db, $table: table),
+          createComputedFieldComposer: () =>
+              $$HabitsTagEntriesTableAnnotationComposer($db: db, $table: table),
+          updateCompanionCallback:
+              ({
+                Value<String> habit = const Value.absent(),
+                Value<String> tag = const Value.absent(),
+                Value<int> rowid = const Value.absent(),
+              }) => HabitsTagEntriesCompanion(
+                habit: habit,
+                tag: tag,
+                rowid: rowid,
+              ),
+          createCompanionCallback:
+              ({
+                required String habit,
+                required String tag,
+                Value<int> rowid = const Value.absent(),
+              }) => HabitsTagEntriesCompanion.insert(
+                habit: habit,
+                tag: tag,
+                rowid: rowid,
+              ),
+          withReferenceMapper: (p0) => p0
+              .map(
+                (e) => (
+                  e.readTable(table),
+                  $$HabitsTagEntriesTableReferences(db, table, e),
+                ),
+              )
+              .toList(),
+          prefetchHooksCallback: ({habit = false, tag = false}) {
+            return PrefetchHooks(
+              db: db,
+              explicitlyWatchedTables: [],
+              addJoins:
+                  <
+                    T extends TableManagerState<
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic,
+                      dynamic
+                    >
+                  >(state) {
+                    if (habit) {
+                      state =
+                          state.withJoin(
+                                currentTable: table,
+                                currentColumn: table.habit,
+                                referencedTable:
+                                    $$HabitsTagEntriesTableReferences
+                                        ._habitTable(db),
+                                referencedColumn:
+                                    $$HabitsTagEntriesTableReferences
+                                        ._habitTable(db)
+                                        .id,
+                              )
+                              as T;
+                    }
+                    if (tag) {
+                      state =
+                          state.withJoin(
+                                currentTable: table,
+                                currentColumn: table.tag,
+                                referencedTable:
+                                    $$HabitsTagEntriesTableReferences._tagTable(
+                                      db,
+                                    ),
+                                referencedColumn:
+                                    $$HabitsTagEntriesTableReferences
+                                        ._tagTable(db)
+                                        .id,
+                              )
+                              as T;
+                    }
+
+                    return state;
+                  },
+              getPrefetchedDataCallback: (items) async {
+                return [];
+              },
+            );
+          },
+        ),
+      );
+}
+
+typedef $$HabitsTagEntriesTableProcessedTableManager =
+    ProcessedTableManager<
+      _$NotZeroDatabase,
+      $HabitsTagEntriesTable,
+      HabitsTagEntry,
+      $$HabitsTagEntriesTableFilterComposer,
+      $$HabitsTagEntriesTableOrderingComposer,
+      $$HabitsTagEntriesTableAnnotationComposer,
+      $$HabitsTagEntriesTableCreateCompanionBuilder,
+      $$HabitsTagEntriesTableUpdateCompanionBuilder,
+      (HabitsTagEntry, $$HabitsTagEntriesTableReferences),
+      HabitsTagEntry,
+      PrefetchHooks Function({bool habit, bool tag})
+    >;
 
 class $NotZeroDatabaseManager {
   final _$NotZeroDatabase _db;
@@ -2000,4 +4080,10 @@ class $NotZeroDatabaseManager {
       $$TagsTableTableTableManager(_db, _db.tagsTable);
   $$TasksTagEntriesTableTableManager get tasksTagEntries =>
       $$TasksTagEntriesTableTableManager(_db, _db.tasksTagEntries);
+  $$HabitsTableTableTableManager get habitsTable =>
+      $$HabitsTableTableTableManager(_db, _db.habitsTable);
+  $$HabitCompletionsTableTableTableManager get habitCompletionsTable =>
+      $$HabitCompletionsTableTableTableManager(_db, _db.habitCompletionsTable);
+  $$HabitsTagEntriesTableTableManager get habitsTagEntries =>
+      $$HabitsTagEntriesTableTableManager(_db, _db.habitsTagEntries);
 }
