@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 part of 'migrations.dart';
 
 Future<void> _from4To5(Migrator m, Schema5 schema) {
@@ -7,7 +23,7 @@ Future<void> _from4To5(Migrator m, Schema5 schema) {
       columnTransformer: {
         schema.tasksTable.forDate: schema.tasksTable.forDate
             .dartCast<DateTime>()
-            .modify(const DateTimeModifier.localTime())
+            .modify(const .localTime())
             .date,
       },
     ),
