@@ -1,4 +1,21 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:flutter/material.dart';
+import 'package:nz_flutter_core/nz_flutter_core.dart';
 
 class ClickableCard extends StatelessWidget {
   const ClickableCard({required this.child, super.key});
@@ -8,9 +25,9 @@ class ClickableCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
-      color: Theme.of(context).cardTheme.color,
+      color: context.theme.cardTheme.color,
       elevation: 2,
-      clipBehavior: Clip.antiAlias,
+      clipBehavior: .antiAlias,
       child: child,
     );
   }
