@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 // coverage:ignore-file
 import 'dart:async';
 
@@ -46,8 +62,10 @@ class _ElevatedCopyButtonState extends State<ElevatedCopyButton> {
         child: AnimatedSwitcher(
           duration: const Duration(milliseconds: 500),
           child: _isCopied
-              ? Text(widget.copiedText ?? t.common.copy.completedText)
-              : Text(widget.defaultCopyText ?? t.common.copy.defaultText),
+              ? Text(widget.copiedText ?? context.t.common.copy.completedText)
+              : Text(
+                  widget.defaultCopyText ?? context.t.common.copy.defaultText,
+                ),
         ),
       ),
     );
