@@ -1,3 +1,19 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'package:flutter/material.dart';
 
 class StarsRateWidget extends StatefulWidget {
@@ -35,12 +51,12 @@ class _StarsRateWidgetState extends State<StarsRateWidget> {
   @override
   Widget build(BuildContext context) {
     return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
+      mainAxisAlignment: .center,
       children: List<Widget>.generate(widget.count, (index) {
         final active = index <= currentValue;
 
         return Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 3),
+          padding: const .symmetric(horizontal: 3),
           child: _StarIcon(
             active: active,
             onTap: () {
