@@ -1,26 +1,44 @@
+// Not Zero, cross-platform wellbeing application.
+// Copyright (C) 2025 Nagorny Vladislav
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:nz_common/nz_common.dart';
 import 'package:nz_flutter_core/src/translations/translations.g.dart';
 
 abstract final class NzDateTimeFormat {
   static String relativeLocalFormat(
+    BuildContext context,
     DateTime dateTime, {
     DateTime? dateToCompare,
     dynamic locale,
   }) {
-    dateToCompare ??= DateTime.now();
+    dateToCompare ??= .now();
 
     // If today
     if (dateTime.isAtSameDay(dateToCompare)) {
-      return t.common.timeOptions.today;
+      return context.t.common.timeOptions.today;
     }
     // If yesterday
     if (dateTime.isAtSameDay(dateToCompare.dayBefore)) {
-      return t.common.timeOptions.yesterday;
+      return context.t.common.timeOptions.yesterday;
     }
     // If tomorrow
     if (dateTime.isAtSameDay(dateToCompare.dayAfter)) {
-      return t.common.timeOptions.tomorrow;
+      return context.t.common.timeOptions.tomorrow;
     }
 
     return localFormat(dateTime, dateToCompare: dateToCompare, locale: locale);
@@ -31,7 +49,7 @@ abstract final class NzDateTimeFormat {
     DateTime? dateToCompare,
     dynamic locale,
   }) {
-    dateToCompare ??= DateTime.now();
+    dateToCompare ??= .now();
 
     // If the same day
     if (dateTime.isAtSameDay(dateToCompare)) {
